ai-service/requirements.txt
flask==3.0.2
pdfminer.six==20231228
python-dotenv==1.0.1
groq
